fix wallpaper script; ranger
Fix wallpaper script so that $1 is being used as the dir.  Remove ranger
bookmarks from git since this will likely be machine specific.  Revert
the column sizes in ranger to their sane defaults so that the columns
can be read.  Removed the custom script from i3status because I'm not
sure how I want to customize it yet.