add script to enhance workspaces
This script allows the user to open a named workspace and open a given applicaton
in that workspace.  If the workspace already exist it is assumed that
the applicatino is already running, and does not attempt to start the application.
If the workspace does not exist it is created and the application is started in the
new workspace.