update config to use named-space & fix named-space
Forgot to update named-space when moving the workspace script. Took the
time to update the named-space script to be more DRY and to pass two
arguments to the workspace script.  This allows the user to create a new
named workspace and open an application in the new workspace by typing a
name and an command into dmenu.